Recursively unset a property
Imagine that we have a deeply nested property
`users/userId/permissions/writeAcces` = `true`.  Then we call remove on
that permission.  `unset` will remove the `writeAccess` property.
However, the `permissions` property will remain on the `userId` object,
*regardless* of if there are any other properties on that object.  In
the case that there was no other property in `permissions`, there will
be an empty `{}` in redux-state for permissions.  *However*, in
firebase - `permissions` will not even exist.  By recursively unsetting
the property, we make sure the parent properties are also removed if
necessary.