Update populate.md
I believe the result would not filter down to just "email", or at least the example doesn't seem to indicate how that would occur.

I also removed the "equivalent string notation" as I can't find what that would be for childAlias.  (Is there documentation of the string notation somewhere?  It seems to be embedded into the Object notation/API documentation?)